% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdImpute.R
\name{flatten_mat}
\alias{flatten_mat}
\title{Flatten and arrange cor matrix to be df}
\usage{
flatten_mat(cor_mat, return_mat)
}
\arguments{
\item{cor_mat}{A correlation matrix output from running \code{feature_cor()}}

\item{return_mat}{Logical. Should the flattened matrix be printed? Default set to FALSE.}
}
\value{
A vector of correlation-based ranked features
}
\description{
Flatten and arrange cor matrix to be df
}
\examples{
\dontrun{
flatten_mat(cor_mat = cor_mat, return_mat = FALSE)
}
}
