% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdImpute.R
\name{feature_cor}
\alias{feature_cor}
\title{High dimensional imputation via batch processed chained random forests
Build correlation matrix}
\usage{
feature_cor(data, return_cor)
}
\arguments{
\item{data}{A data object.}

\item{return_cor}{Logical. Should the correlation matrix be printed? Default set to FALSE.}
}
\value{
A cross-feature correlation matrix
}
\description{
High dimensional imputation via batch processed chained random forests
Build correlation matrix
}
\examples{
\dontrun{
feature_cor(data = data, return_cor = FALSE)
}
}
\references{
van Buuren S, Groothuis-Oudshoorn K (2011). "mice: Multivariate Imputation by Chained Equations in R." Journal of Statistical Software, 45(3), 1-67. doi: <10.18637/jss.v045.i03>
}
